fix: utils.py
